change offscreen vram pixel to 8bit

+++ ssos/os/window/layer.h
@@ -13,7 +13,7 @@ typedef struct {
     uint16_t x, y, z;
     uint16_t w, h;
     uint16_t attr;
-    uint16_t* vram;
+    uint8_t* vram;
 } Layer;
 
 typedef struct {
@@ -27,13 +27,13 @@ extern LayerMgr* ss_layer_mgr;
 
 void ss_layer_init();
 Layer* ss_layer_get();
-void ss_layer_set(Layer* layer, uint16_t* vram, uint16_t x, uint16_t y,
+void ss_layer_set(Layer* layer, uint8_t* vram, uint16_t x, uint16_t y,
                   uint16_t w, uint16_t h);
 void ss_layer_set_z(Layer* layer, uint16_t z);
 void ss_all_layer_draw();
 void ss_all_layer_draw_rect(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
-void ss_layer_draw_rect_layer(Layer *l, uint16_t x0, uint16_t y0, uint16_t x1,
-                          uint16_t y1);
+void ss_layer_draw_rect_layer(Layer* l, uint16_t x0, uint16_t y0, uint16_t x1,
+                              uint16_t y1);
 void ss_layer_move(Layer* layer, uint16_t x, uint16_t y);
 void ss_layer_invalidate(Layer* layer);
-void ss_layer_update_map(Layer *layer);
+void ss_layer_update_map(Layer* layer);
